chore: Объединил Avatar
